tree-wide: Fix implicit casts
These were found by buiding the unix coverage variant on macOS (so
clang compiler). Mostly, these are fixing implicit cast of float/double
to mp_float_t which is one of those two and one mp_int_t to size_t fix
for good measure.

+++ py/objarray.c
@@ -445,7 +445,7 @@ STATIC mp_obj_t array_subscr(mp_obj_t self_in, mp_obj_t index_in, mp_obj_t value
                 }
                 #endif
                 if (len_adj > 0) {
-                    if (len_adj > o->free) {
+                    if ((size_t)len_adj > o->free) {
                         // TODO: alloc policy; at the moment we go conservative
                         o->items = m_renew(byte, o->items, (o->len + o->free) * item_sz, (o->len + len_adj) * item_sz);
                         o->free = len_adj;
